requirements.txt
PyMuPDF
requests
PyPDF2
pillow
pytesseract